git-svn-id: https://www.seasar.org/svn/teeda/trunk@2272 e0972215-f904-0410-97d3-99f68b247671

+++ teeda-html-example/src/main/java/examples/teeda/web/select/SelectOne2Page.java
@@ -11,7 +11,7 @@ public class SelectOne2Page {
 
 	// public static final String aaa_TRequiredValidator = null;
 
-	private Integer aaa = new Integer(1);
+	private Integer aaa = new Integer(2);
 
 	public String prerender() {
 		aaaItems = new ArrayList();
